+++ src/app/admin-view/admin-dashboard/admin-dashboard.component.ts
@@ -1,0 +1,8 @@
+import { ChangeDetectionStrategy, Component } from "@angular/core";
+
+@Component({
+    selector: "app-admin-dashboard",
+    templateUrl: "./admin-dashboard.component.html",
+    changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class AdminDashboardComponent {}