+++ src/app/project-links.ts
@@ -11,6 +11,7 @@ export default class ProjectLinks {
     public static readonly URL_PAGE_NOT_FOUND: string = "v/page-not-found";
     public static readonly URL_RETURN_TO_CLI: string = "v/return-to-cli";
     public static readonly URL_SETTINGS: string = "v/settings";
+    public static readonly URL_ADMIN_DASHBOARD: string = "v/admin-dashboard";
     public static readonly URL_FLOW_DETAILS: string = "flow-details";
 
     public static readonly ALL_URLS: string[] = [